Get Konsumen Sesuai dengan Periode
menjumlahkan nilai fuel jika periode adalah bulanan atau tahunan. terdapat juga get master data.

+++ app/src/main/java/com/pertamina/pertaminatuban/distribusi/models/Konsumen.java
@@ -52,11 +52,11 @@ public class Konsumen {
         this.fuel = fuel;
     }
 
-    public void setNilai(int nilai) {
+    public void setNilai(float nilai) {
         this.nilai = nilai;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(float total) {
         this.total = total;
     }
